requirements.txt
numpy
python-snap7
bitarray